Gardening
Move logger, default_logger method to Delogger class
