+++ delogger/logger.py
@@ -1,14 +1,67 @@
 import atexit
 from copy import copy
+from logging import DEBUG, INFO, WARNING
 from logging.handlers import QueueHandler, QueueListener
 from queue import Queue
 
+from delogger import RunRotatingHandler
 from delogger.base import DeloggerBase
 from delogger.decorators import DeloggerDecorators
 
 
 class Delogger(DeloggerDecorators, DeloggerBase):
-    pass
+    @property
+    def logger(self):
+        """Return set logging.Logger."""
+
+        # Set only loggers that have not been set yet.
+        if self._is_new_logger and self.default:
+            self.default_logger()
+            self._logger.propagate = False
+            self._is_new_logger = False
+
+        return self._logger
+
+    def default_logger(self):
+        """Set default handler."""
+
+        if not self.is_debug_stream and self.stream_level <= INFO:
+            # info is a normal stream, over warning it is a debug stream.
+            fmt = self.stream_fmt
+            self.add_stream_handler(
+                self.stream_level,
+                fmt=fmt,
+                is_color_stream=self.is_color_stream,
+                only_level=True,
+            )
+
+            # Set warning or more stream
+            fmt = self._stream_fmt(is_debug_stream=True)
+            self.add_stream_handler(
+                WARNING, fmt=fmt, is_color_stream=self.is_color_stream
+            )
+
+        else:
+
+            stream_level = self.stream_level
+            if self.is_debug_stream and self.stream_level == INFO:
+                stream_level = DEBUG
+
+            # all debug stream.
+            fmt = self.stream_fmt
+            self.add_stream_handler(
+                stream_level, fmt=fmt, is_color_stream=self.is_color_stream
+            )
+        # If there is a log save flag, the log file handler is set.
+
+        if self.is_save_file:
+            rrh = RunRotatingHandler(
+                self.dirpath,
+                backup_count=self.backup_count,
+                filepath=self.filepath,
+                fmt=self.filename,
+            )
+            self.add_handler(rrh, DEBUG, fmt=self.file_fmt)
 
 
 class DeloggerQueue(Delogger):
